requirements.txt
gradio~=3.28.1
opencv-python~=4.7.0.72
numpy~=1.24.3
Pillow~=9.5.0
onnxruntime~=1.14.1
pandas~=2.0.1
httpx~=0.24.0
requests~=2.29.0
beautifulsoup4~=4.12.2
tqdm~=4.65.0
p_tqdm
h2